+++ CHANGELOG.md
@@ -31,6 +31,7 @@
  - Add Input: INA219 current sensor
  - Add Input: Grove Pi DHT11/22 sensor
  - Add Input: HC-SR04 Ultrasonic Distance sensor
+ - Add Input: SCD30 CO2/Humidity/Temperature sensor
  - Add LCD: Grove RGB LCD
  - Add Function: bang-bang/hysteretic
  - Add Function Action: Output Value